More collisions, got impulse working

+++ DocumentationApp/Body.cpp
@@ -34,7 +34,7 @@ glm::vec2 Body::getVelocity() {
 	return velocity;
 }
 
-double Body::getMass() {
+float Body::getMass() {
 	return mass;
 }
